Fetches the user's lookup_users data then updates their info in the graph.

Fetches the user's lookup_users data then updates their info in the graph.

Usage

add_lookup_users_info_to_nodes_in_graph(
  users,
  token,
  retryonratelimit,
  verbose,
  cache
)

Arguments

users

User id or screen name of target user.

token

Expert use only. Use this to override authentication for a single API call. In most cases you are better off changing the default for all calls. See auth_as() for details.

retryonratelimit

If TRUE, and a rate limit is exhausted, will wait until it refreshes. Most Twitter rate limits refresh every 15 minutes. If FALSE, and the rate limit is exceeded, the function will terminate early with a warning; you'll still get back all results received up to that point. The default value, NULL, consults the option rtweet.retryonratelimit so that you can globally set it to TRUE, if desired.

If you expect a query to take hours or days to perform, you should not rely soley on retryonratelimit because it does not handle other common failure modes like temporarily losing your internet connection.

verbose

Show progress bars and other messages indicating current progress?

cache

A cache object. See nc_create_cache().

Value

tibble of user data