Fetches the friends of each of the users contained in the user_ids vector. Returns the friend relationships via a tibble edge list.

Fetches the friends of each of the users contained in the user_ids vector. Returns the friend relationships via a tibble edge list.

nc_get_friends(
  users,
  cache_name,
  n = 5000,
  retryonratelimit = NULL,
  cursor = "-1",
  verbose = TRUE,
  token = NULL
)

Arguments

users
Screen name or user ID of target user from which the user IDs of friends (accounts followed BY target user) will be retrieved.

n
Desired number of results to return. Results are downloaded in pages when n is large; the default value will download a single page. Set n = Inf to download as many results as possible.

The Twitter API rate limits the number of requests you can perform in each 15 minute period. The easiest way to download more than that is to use retryonratelimit = TRUE.

You are not guaranteed to get exactly n results back. You will get fewer results when tweets have been deleted or if you hit a rate limit. You will get more results if you ask for a number of tweets that's not a multiple of page size, e.g. if you request n = 150 and the page size is 200, you'll get 200 results back.

retryonratelimit
If TRUE, and a rate limit is exhausted, will wait until it refreshes. Most twitter rate limits refresh every 15 minutes. If FALSE, and the rate limit is exceeded, the function will terminate early with a warning; you'll still get back all results received up to that point. The default value, NULL, consults the option rtweet.retryonratelimit so that you can globally set it to TRUE, if desired.

If you expect a query to take hours or days to perform, you should not rely soley on retryonratelimit because it does not handle other common failure modes like temporarily losing your internet connection.

cursor
Which page of results to return. The default will return the first page; you can supply the result from a previous call to continue pagination from where it left off.

verbose
Show progress bars and other messages indicating current progress?

token
Expert use only. Use this to override authentication for a single API call. In most cases you are better off changing the default for all calls. See auth_as() for details.

cache
The name of the cache to save data in.

Value

A tibble where each row corresponds to a follower relationship from the user in the 'from' column to the user in to 'to' column