Get (cached) information about Twitter users

Looks up information pertaining to a vector of provided users. Retrieves cached data for users who have already been looked up, and fetches new data for users who have not bee looked up yet from the Twitter API.

nc_lookup_users(
  users,
  cache_name,
  token = NULL,
  retryonratelimit = TRUE,
  verbose = TRUE
)

Arguments

users

Character vector of Twitter user ids. DO NOT pass Twitter screen names, as users are taken to be unique node identifiers.

token

Expert use only. Use this to override authentication for a single API call. In most cases you are better off changing the default for all calls. See auth_as() for details.

retryonratelimit

If TRUE, and a rate limit is exhausted, will wait until it refreshes. Most Twitter rate limits refresh every 15 minutes. If FALSE, and the rate limit is exceeded, the function will terminate early with a warning; you'll still get back all results received up to that point. The default value, NULL, consults the option rtweet.retryonratelimit so that you can globally set it to TRUE, if desired.

If you expect a query to take hours or days to perform, you should not rely soley on retryonratelimit because it does not handle other common failure modes like temporarily losing your internet connection.

verbose

Show progress bars and other messages indicating current progress?

cache

A cache object. See nc_create_cache().

Value

A tibble where each row corresponds to a User and each column to one of the User properties. If a user cannot be sampled, should return nothing for that user. If no users can be sampled, should return an empty tibble with appropriate columns.